refactor: consolidate Orbit classes
