+++ src/galax/dynamics/__init__.py
@@ -5,9 +5,7 @@ __all__ = [
     "integrate",
     "mockstream",
     # orbit, et al.
-    "AbstractOrbit",
     "Orbit",
-    "InterpolatedOrbit",
     # integrate
     "evaluate_orbit",
     # mockstream
@@ -31,7 +29,6 @@ from galax.setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("galax.dynamics", RUNTIME_TYPECHECKER):
     from . import integrate, mockstream
-    from ._dynamics.base import AbstractOrbit
     from ._dynamics.funcs import (
         lagrange_points,
         specific_angular_momentum,
@@ -40,7 +37,7 @@ with install_import_hook("galax.dynamics", RUNTIME_TYPECHECKER):
     from ._dynamics.integrate.funcs import evaluate_orbit
     from ._dynamics.mockstream import MockStream, MockStreamArm, MockStreamGenerator
     from ._dynamics.mockstream.df import AbstractStreamDF, ChenStreamDF, FardalStreamDF
-    from ._dynamics.orbit import InterpolatedOrbit, Orbit
+    from ._dynamics.orbit import Orbit
 
 
 # Cleanup
