+++ tests/unit/dynamics/test_orbit.py
@@ -42,12 +42,12 @@ class TestOrbit(AbstractPhaseSpacePosition_Test[Orbit]):
         q = Quantity(jr.normal(next(subkeys), (*shape, 3)), "kpc")
         p = Quantity(jr.normal(next(subkeys), (*shape, 3)), "km/s")
         t = Quantity(jr.normal(next(subkeys), shape[-1:]), unit=potential.units["time"])
-        return w_cls(q=q, p=p, t=t, potential=potential)
+        return w_cls(q=q, p=p, t=t, potential=potential, interpolant=None)
 
     @pytest.fixture
     def w(self, w_cls: type[T], shape: gt.Shape, potential: AbstractPotentialBase) -> T:
         """Return a phase-space position."""
-        return self.make_w(w_cls, shape, potential)
+        return self.make_w(w_cls, shape, potential=potential)
 
     # ===============================================================
